Add consignment and view consignment done

+++ lib/models/routes_for_client_id_response.dart
@@ -87,7 +87,7 @@ class Hubs {
   final int id;
   final int hub;
   final int sequence;
-  final int kms;
+  final double kms;
   final int tag;
 
   Hubs copyWith({
